+++ CHANGELOG.md
@@ -45,6 +45,7 @@
 - SyncingBlockEntity
 - Label offsets in BaseScreen
 - AaronClientUtil#getBiomeDisplay
+- ItemStackButton now uses a Supplier<ItemStack> instead of just an ItemStack
 
 # 1.6.0